Only test xor4096i, and show first 5 random numbers.

+++ doc/xorgens/xorgens.h
@@ -1,10 +1,11 @@
 /* xorgens.h for xorgens version 3.04, R. P. Brent 20060628 */
+#include <stdint.h>
 
-typedef unsigned long UINT; /* Type for random 32 or 64-bit integer, 
+typedef uint32_t UINT; /* Type for random 32 or 64-bit integer, 
                                e.g. unsigned long, unsigned long long,
                                uint64_t, unsigned int or uint32_t */
                                    
-typedef double UREAL;       /* Type for random 32 or 64-bit real,
+typedef float UREAL;       /* Type for random 32 or 64-bit real,
                                e.g. double or float */
 
 UINT  xor4096i(UINT seed);  /* integer random number generator */
